implemented bcp analysis

+++ analyse/helpers.go
@@ -2,11 +2,11 @@ package analyse
 
 import "bitbucket.org/sealuzh/gopper/data"
 
-func vectorise(r *data.TestResult) []float32 {
+func vectorise(r *data.TestResult) []float64 {
 	l := len(r.ExecutionResults)
-	ret := make([]float32, l)
+	ret := make([]float64, l)
 	for i, r := range r.ExecutionResults {
-		ret[i] = r.RawVal
+		ret[i] = float64(r.RawVal)
 	}
 	return ret
 }
